tools/consensus_cli/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "consensus-cli"
version = "0.1.0"
description = "Multi-LLM consensus CLI (Gemini, OpenAI, Claude, Ollama)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "litellm>=1.55.0",
  "typer>=0.12.0",
  "rich>=13.0.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
book-consensus = "consensus_cli.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["consensus_cli*"]
